Decompose job expansion into one function per file
expandJobs keeps the loop; the callee-input resolution, granted execution,
reusable-workflow resolution, uses expansion, and per-job verdict each move
to their own module under src/jobs/.
